requirements.txt
altgraph==0.17
future==0.18.2
lxml==4.5.2
pefile==2019.4.18
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
python-docx==0.8.10
pywin32-ctypes==0.2.0
